pattern match update, should be done
